Started algorithm implementation

+++ AlgoLogistics/AlgoLogistics.Algorithms/Dijkstra/DijkstraAlgorithmInput.cs
@@ -7,7 +7,7 @@ namespace AlgoLogistics.Algorithms.Dijkstra
 {
 	public class DijkstraAlgorithmInput
 	{
-		public Dictionary<string, Dictionary<string, int>> Graph { get; set; }
+		public Dictionary<string, Dictionary<string, double>> Graph { get; set; }
 		public string Start { get; set; }
 		public string End { get; set; }
 	}
